+++ SeriesGuide/src/main/java/com/battlelancer/seriesguide/tmdbapi/SgTmdb.java
@@ -33,6 +33,6 @@ public class SgTmdb extends Tmdb {
     @Override
     protected RestAdapter.Builder newRestAdapterBuilder() {
         return new RestAdapter.Builder().setClient(
-                new OkClient(ServiceUtils.getOkHttpClient(context)));
+                new OkClient(ServiceUtils.getCachingOkHttpClient(context)));
     }
 }