+++ CHANGELOG.md
@@ -8,6 +8,14 @@ All dates are in the European Central timezone.
 Version 17 *(in development)*
 -----------------------------
 
+#### 17-beta4 *(2014-08-24)*
+
+* TWEAK Pressing back when third-party app linked to show or episode returns to third-party app (see https://github.com/UweTrottmann/SeriesGuide/wiki/Extension-API).
+* TWEAK Accept stale images when offline. This should improve images displaying if offline for longer periods of time.
+* FIX Potential fix for image cache getting corrupted. Also, images are now exclusively cached on internal storage (avoids external storage format issues).
+* FIX Small UI fixes.
+* NOTICE Latest translations from crowdin.
+
 #### 17-beta3 *(2014-08-15)*
 
 * TWEAK Remove show dialog opens without delay.